fixing memcpy

+++ ips/memcpy/hw/src/memcpy_top.hpp
@@ -17,7 +17,7 @@
 //#include "ap_int.h"
 
 typedef uint32_t args_t;
-typedef uint32_t data_t;
+typedef uint64_t data_t;
 
 static const uint8_t ARGS_SIZE = 8;
